+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - discogs: Improved search for Draft release in popup
 - bandcamp: Auto-detection of the digital release audio quality
 
+### Fixed
+
+- bandcamp: Fixed the issue with the artist/album filtering on the artist page
+
 ## 0.19.0 (2025-01-28)
 
 ### Features